feat(feedbacks.tsx): Refactor Feedbacks component to use FeedbackCard component The Feedbacks component has been refactored to use the FeedbackCard component. This improves code organization and reusability by separating the logic for rendering individual feedback cards into a separate component.
refactor(tech.tsx): remove unused imports and interfaces, move TechnologiesCard component to separate file
refactor(blog.tsx): remove unused imports, move BlogPostCard component to separate file
refactor(experience.tsx): remove unused imports, move ExperienceCard component to separate file
refactor(work.tsx): remove unused imports, move ProjectCard component to separate file

The changes were made to improve code organization and maintainability. By moving the components to separate files, the code becomes more modular and easier to understand. Removing unused imports helps to reduce clutter and improve performance.
